pyproject.toml
[project]
name = "umamusume-agent-prompt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crawl4ai>=0.7.8",
    "google-api-python-client>=2.127.0",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "langchain>=1.2.0",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=0.3.0",
    "langgraph>=0.2.0",
    "mcp>=1.25.0",
    "mcp-server>=0.1.4",
    "mutagen>=1.47.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24.0",
    "python-dotenv>=1.0.1",
    "starlette>=0.37.2",
    "tqdm>=4.67.1",
    "uvicorn>=0.30.0",
]
